requirements.txt
opencv-python==4.7.0.68
pandas==1.5.3
gradio==3.16.2
scikit-learn==1.2.1
scikit-image==0.19.3
Pillow==9.4.0
tqdm==4.63.0
diffusers==0.27.2
gradio==3.16.2
gradio_client==0.2.5

